Separate schemas/models into domains
