+++ src/applications/schemas.py
@@ -1,0 +1,4 @@
+from src.applications.models import ApplicationBase
+
+class ApplicationCreateRequest(ApplicationBase):
+    pass
