+++ src/database/mongo/core.py
@@ -5,7 +5,7 @@ from pymongo.server_api import ServerApi
 from bson import json_util
 
 from src.config import MONGO_DATABASE_NAME
-from src.database.mongo.services import init_collections
+from src.database.mongo.service import init_collections
 
 MONGO_URL = environ.get("CTI_MONGO_URL")
 if not MONGO_URL:
